Patterns

+++ src/patterns/butterfly/ButterflyShape.java
@@ -2,12 +2,12 @@ package patterns.butterfly;
 
 public class ButterflyShape {
     public void ButterFly(int row){
-        for(int i=1; i<=row; i++){
-            for(int j=1; j<=i; j++){
+        for(byte i=1; i<=row; i++){
+            for(byte j=1; j<=i; j++){
                 System.out.print("*");
             }
             int space = 2*(row-i);
-            for(int j=1; j<=space; j++){
+            for(short j=1; j<=space; j++){
                 System.out.print(" ");
             }
             for(int j=1; j<=i; j++){
